test-lib: drop comment about test_description
When a comment describing how each test file should start was added in
commit [1], it was the second comment of t/test-lib.sh.  The comment
describes how variable "test_description" is supposed to be assigned at
the top of each test file.  However, even in [1], the comment was ten
lines away from the usage of the variable by test-lib.sh.  Since then,
the comment has drifted away both from the top of the file and from the
usage of the variable.  The comment just sits in the middle of the
initialization of the test library, surrounded by unrelated code, almost
one hundred lines away from the usage of "test_description".

Nobody has noticed this drift during evolution of test-lib.sh, which
suggests that this comment has outlived its usefulness.  The assignment
of "test_description" and the process of writing tests in general are
described in detail in "t/README".  So drop the obsolete comment.

An alternative solution is to move the comment down to the usage of
variable "test_description".

[1] e1970ce43a ("[PATCH 1/2] Test framework take two.", 2005-05-13)

Signed-off-by: Andrei Rybak <rybak.a.v@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>